requirements.txt
flask
google-generativeai
moviepy
python-docx 
markdown
docx
beautifulsoup4